f: parallel routes behavior for asset detail
