+++ apps/wallet/src/routes/portfolio/assets/index.tsx
@@ -1,8 +1,8 @@
 import { AssetsList } from '@status-im/wallet/components'
-import { useQuery } from '@tanstack/react-query'
 import { createFileRoute, useRouter } from '@tanstack/react-router'
 
 import SplittedLayout from '@/components/splitted-layout'
+import { useAssets } from '@/hooks/use-assets'
 
 export const Route = createFileRoute('/portfolio/assets/')({
   component: Component,
@@ -10,57 +10,7 @@ export const Route = createFileRoute('/portfolio/assets/')({
 
 function Component() {
   const router = useRouter()
-
-  // const handleSelect = (url: string, options?: { scroll?: boolean }) => {
-  //   // Handle the selection of an asset
-  //   console.log('Selected asset URL:', url)
-  //   console.log('Scroll option:', options?.scroll)
-  // }
-
-  // todo: export trpc client with api router and used instead
-  // todo: cache
-  const { data: assets, isLoading } = useQuery({
-    queryKey: ['assets'],
-    queryFn: async () => {
-      const url = new URL('http://localhost:3030/api/trpc/assets.all')
-      url.searchParams.set(
-        'input',
-        JSON.stringify({
-          json: {
-            address: '0xd8da6bf26964af9d7eed9e03e53415d37aa96045',
-            networks: [
-              'ethereum',
-              'optimism',
-              'arbitrum',
-              'base',
-              'polygon',
-              'bsc',
-            ],
-          },
-        }),
-      )
-
-      const response = await fetch(url, {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      })
-
-      if (!response.ok) {
-        throw new Error('Failed to fetch.')
-      }
-
-      const body = await response.json()
-
-      return body.result.data.json.assets
-    },
-    staleTime: 60 * 60 * 1000, // 1 hour
-    gcTime: 60 * 60 * 1000, // 1 hour
-    refetchOnMount: false,
-    refetchOnWindowFocus: false,
-    refetchOnReconnect: false,
-  })
+  const { data: assets, isLoading } = useAssets()
 
   return (
     <SplittedLayout
